Refactoring a high level interface for different meshing target systems, refactoring the spherical system to be a generic implementation
